tests: make application testable
* Refactors application instantiation to a factory so it is
  easily testable.

* Adds application and mocked job fixture to test all scenarios.

* Introduce custom exception in order to have an abstract way of defining
  computing backend errors.
