+++ docs/restapi.rst
@@ -6,6 +6,6 @@ REST API
 The REANA Job Controller API offers different endpoints to create, manage and monitor jobs.
 Detailed REST API documentation can be found `here <_static/api.html>`_.
 
-.. automodule:: reana_job_controller.app
+.. automodule:: reana_job_controller.rest
    :members:
    :exclude-members: get_openapi_spec
