+++ reana_job_controller/errors.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of REANA.
+# Copyright (C) 2018 CERN.
+#
+# REANA is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""REANA-Job-Controller errors."""
+
+
+class ComputingBackendSubmissionError(Exception):
+    """Operation to computer backend could not be performed."""
